observableFrom eagerly consumes to avoid lazy CP gotcha
If you're using observableFrom('someCp'), it's
likely the Ember observer won't correctly attach
due to the lazy CP optimization in Ember that requires
eager consumption of a computed property if you want
observers to properly attach.

This commit makes observableFrom eagerly consumed
the observed property.

It also bumps ember-cli because I was running into
rim-raf errors that are fixed in 0.2.0